move is_gdf
